+++ src/javascript/API.js
@@ -365,6 +365,28 @@ const graviton = {
       }
     });
   },
+  closingFileWarn : function(){
+    new Dialog({
+      id: "saving_file_warn",
+      title: current_config.language["Warn"],
+      content: current_config.language["FileExit-dialog-message"],
+      buttons: {
+        [current_config.language[
+          "FileExit-dialog-button-accept"
+        ]]: {
+          click:()=>{
+            closeTab(ele.getAttribute("tabid"),true);
+          }
+        },
+        [current_config.language["Cancel"]]: {},
+        [current_config.language["FileExit-dialog-button-deny"]]: {
+          click: ()=>{saveFile()},
+          important: true
+        }
+      }
+    });
+  }
+  ,
   addContextMenu: function(panel) {
     Object.keys(panel).forEach(function(key) {
       context_menu_list_text[key] = panel[key];