+++ src/javascript/top_bar.js
@@ -167,8 +167,6 @@ Help.setList({
     "Source Code": () =>
       shell.openExternal("https://github.com/Graviton-Code-Editor"),
     "Telegram Channel": () => shell.openExternal("https://t.me/gravitoneditor"),
-    "Telegram Group": () =>
-      shell.openExternal("https://t.me/joinchat/FgdqbBRNJjpSHPHuDRMzfQ"),
     "*line": "",
     Donate: () => shell.openExternal("https://www.paypal.me/mkenzo8"),
     Twitter:() => shell.openExternal('https://twitter.com/gravitoneditor'),